requirements.txt
numpy
torch
accelerate
transformers
matplotlib
pandas
datasets
librosa
ipykernel
torchsummary
sounddevice
openai-whisper
